create separate crate for poseidon parameter generation
