+++ poseidon-paramgen/src/matrix.rs
@@ -1,0 +1,4 @@
+use ark_ff::PrimeField;
+
+/// Represents a matrix over Fp
+pub struct Matrix<F: PrimeField>(Vec<Vec<F>>);
