+++ poseidon377/src/lib.rs
@@ -2,8 +2,6 @@
 
 use once_cell::sync::Lazy;
 
-pub mod param_gen;
-
 mod hash;
 mod params;
